# Server/toeictesting247-swagger.yaml
openapi: 3.0.3
info:
  title: Tài liệu API Toeictesting247 - Xây dựng website học tiếng Anh Toeic trực tuyến
  description: |-
    ## Mục Đích

    Tài liệu này được viết để hướng dẫn lập trình viên Front-end có thể hiểu và sử dụng các API của Toeictesting247 một cách dễ dàng. API cung cấp các chức năng quản lý người dùng, quản lý khoá học, quản lý tài liệu, quản lý bài kiểm tra, quản lý câu hỏi, quản lý ảnh, video, audio, và nhiều tính năng khác.

    ## Thông Tin API

    - **Ngôn Ngữ Lập Trình**: TypeScript
    - **Môi Trường**: Node.js
    - **Tài Liệu**: Swagger Editor, Swagger UI

    ## Các Chức Năng Chính

    - **Quản Lý Người Dùng**: API cho phép tạo, sửa đổi và xóa thông tin người dùng.
    - **Quản Lý Khoá Học**: Cung cấp chức năng thêm, sửa đổi, xóa khoá học.
    - **Quản Lý Tài Liệu**: API cho phép quản lý các tài liệu, bao gồm ảnh, video, audio.
    - **Quản Lý Bài Kiểm Tra**: Cung cấp chức năng tạo, sửa đổi, xóa bài kiểm tra.
    - **Quản Lý Câu Hỏi**: Cho phép quản lý các câu hỏi liên quan đến bài kiểm tra.
    - **Quản Lý Multimedia**: API cho phép quản lý ảnh, video, audio.
      
      **...**

    ## Đặc Điểm của Tài Liệu API

    - **Chia nhóm**: API được chia thành các nhóm theo tag, mỗi tag tương ứng với một nhóm API cụ thể.
    - **Mô tả chi tiết**: Mỗi API đều có mô tả chi tiết về input, output, cách sử dụng, cách test.
    - **Ví dụ**: Mỗi API đều có một ví dụ minh họa cách sử dụng.

    ## Lưu Ý

...

  contact:
...
  license:
    name: Apache 2.0
...
servers:
...
paths:
...
  /tests-v2/add-question:
    patch:
      tags:
        - tests_v2
      summary: Thêm câu hỏi vào bài kiểm tra
      description: Thêm câu hỏi vào bài kiểm tra
      operationId: addQuestionToTest
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin thêm câu hỏi vào bài kiểm tra
        content:
          application/json:
            schema:
              type: object
              properties:
                test_id:
                  type: string
                  format: MongoId
                  description: id của bài kiểm tra
                  example: '660b4cab4673a97618935954'
                question_id:
                  type: string
                  format: MongoId
                  description: id của câu hỏi
                  example: '660b4cab4673a97618935954'
        required: true
      responses:
        default:
          description: Thêm câu hỏi vào bài kiểm tra thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Add question to test successfully
                  result:
                    $ref: '#/components/schemas/TestResponseBody_v2'
  /tests-v2/random-questions:
    patch:
      tags:
        - tests_v2
      summary: Thêm câu hỏi vào bài kiểm tra
      description: Thêm câu hỏi vào bài kiểm tra
      operationId: addQuestionToTest
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin thêm câu hỏi vào bài kiểm tra
        content:
          application/json:
            schema:
              type: object
              properties:
                test_id:
                  type: string
                  format: MongoId
                  description: id của bài kiểm tra
                  example: '660b4cab4673a97618935954'
                type_question:
                  $ref: '#/components/schemas/QuestionType'
                type_content_question:
                  $ref: '#/components/schemas/QuestionTypeContent'
                number_question:
                  type: integer
                  example: 2
        required: true
      responses:
        default:
          description: Thêm câu hỏi vào bài kiểm tra thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Add question to test successfully
                  result:
                    $ref: '#/components/schemas/TestResponseBody_v2'
...
components:
  schemas:
...
    TestResponseBody_v2:
      type: object
      properties:
        _id:
          type: string
          example: '6655f6fb91219247ab3fab70'
        title:
          type: string
          example: 'Test 4'
        description:
          type: string
          example: 'Part 1: Photos'
        timeline:
          type: integer
          example: 120
        questions:
          type: array
          example: []
          items:
            type: object
        created_at:
          type: string
          format: date-time
          example: '2024-05-28T15:23:39.603Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-05-28T15:23:39.603Z'
...
    QuestionType:
      type: number
      enum: [SimpleQuestion, QuoteQuestion, DoubleQuestion, TripleQuestion, QuadrupleQuestion, QuintupleQuestion]
      example: 0
    QuestionTypeContent:
      type: number
      enum: [Listening, Reading]
      example: 0
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT